requirements.txt
bs4
beautifulsoup4==4.4.1
colorama==0.3.7
Cython==0.24
Django==1.11
gunicorn==19.7.1
mongoengine==0.13.0
py==1.4.31
pymongo==3.4.0
pytest==2.9.2
pytz==2017.2
requests==2.9.1
six==1.10.0
whitenoise==3.3.0
django-background-tasks==1.1.10
